More consistent use of size_t in place of (u)int/(u)long in gui/textedit files

+++ source/voxelman/gui/textedit/cursor.d
@@ -9,7 +9,7 @@ import voxelman.gui.textedit.linebuffer;
 
 struct Cursor
 {
-	ulong byteOffset;
+	size_t byteOffset;
 	int line;
 }
